% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate model inputs}
\usage{
validate_inputs(csv_data, parameters)
}
\arguments{
\item{csv_data}{A string representing a csv file}

\item{parameters}{A list of parameters}
}
\description{
Validate model inputs
}
